create anim alpha Palette
