+++ app/src/main/res/anim/alpha_out.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:duration="2000">
+    android:duration="350">
     <alpha
         android:fromAlpha="1.0"
         android:toAlpha="0.0" />
